_OFT/requirements.txt
torch==2.8.0
torchvision==0.23.0
diffusers==0.36.0
transformers==4.57.1
accelerate==1.11.0
peft==0.17.1
datasets==4.0.0
safetensors==0.6.2
pillow>=10.0.0
tqdm>=4.66.0
